solved p10.1 and p10.2

+++ src/main/java/it/lpleo/adventofcode/domain/Segment.java
@@ -11,5 +11,5 @@ public class Segment {
 
   private Point a;
   private Point b;
-  private int length;
+  private double length;
 }
